Split base class to all others. Add PopupMessage base class for Interstitial and CenterPopup.
