+++ AndroidSDKCore/src/main/java/com/leanplum/messagetemplates/CenterPopup.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014, Leanplum, Inc. All rights reserved.
+ * Copyright 2020, Leanplum, Inc. All rights reserved.
  *
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -23,25 +23,67 @@ package com.leanplum.messagetemplates;
 
 import android.app.Activity;
 import android.content.Context;
-
+import android.graphics.Point;
+import android.os.Build;
+import android.view.Window;
+import android.view.WindowManager;
+import android.widget.RelativeLayout;
 import com.leanplum.ActionContext;
 import com.leanplum.Leanplum;
 import com.leanplum.LeanplumActivityHelper;
 import com.leanplum.callbacks.ActionCallback;
 import com.leanplum.callbacks.PostponableAction;
 import com.leanplum.callbacks.VariablesChangedCallback;
+import com.leanplum.utils.SizeUtil;
 
 /**
  * Registers a Leanplum action that displays a custom center popup dialog.
  *
  * @author Andrew First
  */
-public class CenterPopup extends BaseMessageDialog {
+public class CenterPopup extends PopupMessageTemplate {
   private static final String NAME = "Center Popup";
 
-  public CenterPopup(Activity activity, CenterPopupOptions options) {
-    super(activity, false, options, null, null);
-    this.options = options;
+  private CenterPopup(Activity activity, CenterPopupOptions options) {
+    super(activity, options, false);//activity, fullscreen=false, options, webOptions=null, htmlOptions=null);
+  }
+
+  @Override
+  protected void initWindowDecoration() {
+    Window window = getWindow();
+    if (window == null) {
+      return;
+    }
+    window.addFlags(WindowManager.LayoutParams.FLAG_DIM_BEHIND);
+    if (Build.VERSION.SDK_INT >= 14) {
+      window.setDimAmount(0.7f);
+    }
+  }
+
+  @Override
+  protected RelativeLayout.LayoutParams createLayoutParams() {
+    RelativeLayout.LayoutParams layoutParams;
+
+    // Make sure the dialog fits on screen.
+    Point size = SizeUtil.getDisplaySize(activity);
+    int width = SizeUtil.dpToPx(activity, ((CenterPopupOptions) options).getWidth());
+    int height = SizeUtil.dpToPx(activity, ((CenterPopupOptions) options).getHeight());
+
+    int maxWidth = size.x - SizeUtil.dp20;
+    int maxHeight = size.y - SizeUtil.dp20;
+    double aspectRatio = width / (double) height;
+    if (width > maxWidth && (int) (width / aspectRatio) < maxHeight) {
+      width = maxWidth;
+      height = (int) (width / aspectRatio);
+    }
+    if (height > maxHeight && (int) (height * aspectRatio) < maxWidth) {
+      height = maxHeight;
+      width = (int) (height * aspectRatio);
+    }
+
+    layoutParams = new RelativeLayout.LayoutParams(width, height);
+    layoutParams.addRule(RelativeLayout.CENTER_IN_PARENT, RelativeLayout.TRUE);
+    return layoutParams;
   }
 
   public static void register(Context currentContext) {
