+++ AndroidSDKCore/src/main/java/com/leanplum/messagetemplates/HTMLTemplate.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017, Leanplum, Inc. All rights reserved.
+ * Copyright 2020, Leanplum, Inc. All rights reserved.
  *
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -21,23 +21,41 @@
 
 package com.leanplum.messagetemplates;
 
+import android.annotation.SuppressLint;
 import android.app.Activity;
-import android.app.AlertDialog;
-import android.content.DialogInterface;
+import android.app.Dialog;
+import android.content.Context;
+import android.graphics.Color;
 import android.graphics.Point;
-import androidx.annotation.NonNull;
+import android.graphics.drawable.ShapeDrawable;
+import android.os.Build;
+import android.os.Handler;
+import android.text.TextUtils;
 import android.view.Gravity;
 import android.view.MotionEvent;
+import android.view.View;
+import android.view.ViewGroup.LayoutParams;
+import android.view.Window;
 import android.view.WindowManager;
-
+import android.webkit.WebChromeClient;
+import android.webkit.WebSettings;
+import android.webkit.WebView;
+import android.webkit.WebViewClient;
+import android.widget.RelativeLayout;
+import androidx.annotation.NonNull;
 import com.leanplum.ActionContext;
 import com.leanplum.Leanplum;
 import com.leanplum.LeanplumActivityHelper;
 import com.leanplum.callbacks.ActionCallback;
 import com.leanplum.callbacks.PostponableAction;
 import com.leanplum.callbacks.VariablesChangedCallback;
+import com.leanplum.core.R;
 import com.leanplum.internal.Log;
 import com.leanplum.utils.SizeUtil;
+import java.io.UnsupportedEncodingException;
+import java.net.URLDecoder;
+import java.util.Map;
+import org.json.JSONObject;
 
 /**
  * Registers a Leanplum action that displays a HTML message.
@@ -48,15 +66,320 @@ import com.leanplum.utils.SizeUtil;
 public class HTMLTemplate extends BaseMessageDialog {
   private static final String NAME = "HTML";
 
-  public HTMLTemplate(Activity activity, HTMLOptions htmlOptions) {
-    super(activity, htmlOptions.isFullScreen(), null, null, htmlOptions);
+  private @NonNull HTMLOptions htmlOptions;
+
+  public HTMLTemplate(Activity activity, @NonNull HTMLOptions htmlOptions) {
+    super(activity);//, htmlOptions.isFullScreen(), null, null, htmlOptions);
     this.htmlOptions = htmlOptions;
+
+    init(activity, htmlOptions.isFullScreen());
+  }
+
+  private void init(Activity activity, boolean fullscreen) {
+
+    SizeUtil.init(activity);
+    dialogView = new RelativeLayout(activity);
+    RelativeLayout.LayoutParams layoutParams = new RelativeLayout.LayoutParams(
+        LayoutParams.MATCH_PARENT, LayoutParams.MATCH_PARENT);
+    dialogView.setBackgroundColor(Color.TRANSPARENT);
+    dialogView.setLayoutParams(layoutParams);
+
+    RelativeLayout view = createContainerView(activity, fullscreen);
+    view.setId(R.id.container_view);
+    dialogView.addView(view, view.getLayoutParams());
+
+    setContentView(dialogView, dialogView.getLayoutParams());
+
+    dialogView.setAnimation(createFadeInAnimation());
+
+    if (!fullscreen) {
+      Window window = getWindow();
+      if (window == null) {
+        return;
+      }
+
+      window.clearFlags(WindowManager.LayoutParams.FLAG_DIM_BEHIND);
+
+      if (htmlOptions.isBannerWithTapOutsideFalse()) {
+        // banners need to be positioned at the top manually
+        // (unless they get repositioned to the bottom later)
+        window.setLayout(LayoutParams.MATCH_PARENT, LayoutParams.WRAP_CONTENT);
+        window.setGravity(Gravity.TOP);
+
+        // use the html y offset to determine the y location of the window; this is different
+        // from non-banners because we don't want to make the window too big (e.g. via a margin
+        // in the layout) and block other things on the screen (e.g. dialogs)
+        WindowManager.LayoutParams windowLayoutParams = window.getAttributes();
+        windowLayoutParams.y = htmlOptions.getHtmlYOffset(activity);
+        window.setAttributes(windowLayoutParams);
+
+        window.setFlags(WindowManager.LayoutParams.FLAG_NOT_TOUCH_MODAL | WindowManager.LayoutParams.FLAG_NOT_FOCUSABLE,
+            WindowManager.LayoutParams.FLAG_NOT_TOUCH_MODAL | WindowManager.LayoutParams.FLAG_NOT_FOCUSABLE);
+      } else {
+        window.setFlags(WindowManager.LayoutParams.FLAG_NOT_TOUCH_MODAL,
+            WindowManager.LayoutParams.FLAG_NOT_TOUCH_MODAL);
+      }
+
+      if (htmlOptions.isHtmlAlignBottom()) {
+        if (htmlOptions.isBannerWithTapOutsideFalse()) {
+          window.setGravity(Gravity.BOTTOM);
+        } else {
+          dialogView.setGravity(Gravity.BOTTOM);
+        }
+      }
+    }
+  }
+
+  @SuppressWarnings("deprecation")
+  private RelativeLayout createContainerView(Activity context, boolean fullscreen) {
+    RelativeLayout view = new RelativeLayout(context);
+
+    // Positions the dialog.
+    RelativeLayout.LayoutParams layoutParams;
+    if (fullscreen) {
+      layoutParams = new RelativeLayout.LayoutParams(
+          LayoutParams.MATCH_PARENT, LayoutParams.MATCH_PARENT);
+    } else {
+      int height = SizeUtil.dpToPx(context, htmlOptions.getHtmlHeight());
+      HTMLOptions.Size htmlWidth = htmlOptions.getHtmlWidth();
+      if (htmlWidth == null || TextUtils.isEmpty(htmlWidth.type)) {
+        layoutParams = new RelativeLayout.LayoutParams(
+            LayoutParams.MATCH_PARENT, height);
+      } else {
+        int width = htmlWidth.value;
+        if ("%".equals(htmlWidth.type)) {
+          Point size = SizeUtil.getDisplaySize(context);
+          width = size.x * width / 100;
+        } else {
+          width = SizeUtil.dpToPx(context, width);
+        }
+        layoutParams = new RelativeLayout.LayoutParams(width, height);
+      }
+
+      layoutParams.addRule(RelativeLayout.CENTER_HORIZONTAL, RelativeLayout.TRUE);
+      int htmlYOffset = htmlOptions.getHtmlYOffset(context);
+      if (!htmlOptions.isBannerWithTapOutsideFalse()) {
+        if (htmlOptions.isHtmlAlignBottom()) {
+          layoutParams.bottomMargin = htmlYOffset;
+        } else {
+          layoutParams.topMargin = htmlYOffset;
+        }
+      }
+    }
+
+    view.setLayoutParams(layoutParams);
+
+    ShapeDrawable footerBackground = new ShapeDrawable();
+    footerBackground.setShape(createRoundRect(fullscreen ? 0 : SizeUtil.dp20));
+    footerBackground.getPaint().setColor(0x00000000);
+    if (Build.VERSION.SDK_INT >= 16) {
+      view.setBackground(footerBackground);
+    } else {
+      view.setBackgroundDrawable(footerBackground);
+    }
+
+    webView = createHtml(context);
+    view.addView(webView, webView.getLayoutParams());
+
+    return view;
+  }
+
+  /**
+   * Create WebView with HTML template.
+   *
+   * @param context Current context.
+   * @return WebVew WebVew with HTML template.
+   */
+  @SuppressLint("SetJavaScriptEnabled")
+  private WebView createHtml(Context context) {
+    dialogView.setVisibility(View.GONE);
+    final WebView webView = new WebView(context);
+    webView.setBackgroundColor(Color.TRANSPARENT);
+    // Disable long click.
+    webView.setLongClickable(false);
+    webView.setHapticFeedbackEnabled(false);
+    webView.setOnLongClickListener(new View.OnLongClickListener() {
+      @Override
+      public boolean onLongClick(View v) {
+        return true;
+      }
+    });
+
+    WebSettings webViewSettings = webView.getSettings();
+    if (Build.VERSION.SDK_INT >= 17) {
+      webViewSettings.setMediaPlaybackRequiresUserGesture(false);
+    }
+    webViewSettings.setAppCacheEnabled(true);
+    webViewSettings.setAllowFileAccess(true);
+    webViewSettings.setJavaScriptEnabled(true);
+    webViewSettings.setDomStorageEnabled(true);
+    webViewSettings.setJavaScriptCanOpenWindowsAutomatically(true);
+    webViewSettings.setLoadWithOverviewMode(true);
+    webViewSettings.setLoadsImagesAutomatically(true);
+
+    if (Build.VERSION.SDK_INT >= 16) {
+      webViewSettings.setAllowFileAccessFromFileURLs(true);
+      webViewSettings.setAllowUniversalAccessFromFileURLs(true);
+    }
+
+    webViewSettings.setBuiltInZoomControls(false);
+    webViewSettings.setDisplayZoomControls(false);
+    webViewSettings.setSupportZoom(false);
+
+    RelativeLayout.LayoutParams layoutParams = new RelativeLayout.LayoutParams(
+        LayoutParams.MATCH_PARENT, LayoutParams.MATCH_PARENT);
+    webView.setLayoutParams(layoutParams);
+    final Dialog currentDialog = this;
+    webView.setWebChromeClient(new WebChromeClient());
+    webView.setWebViewClient(new WebViewClient() {
+      // shouldOverrideUrlLoading(WebView wView, String url) was deprecated at API 24.
+      @SuppressWarnings("deprecation")
+      @Override
+      public boolean shouldOverrideUrlLoading(WebView wView, String url) {
+        if (isClosing) // prevent multiple clicks on same button
+          return true;
+
+        // Open URL event.
+        if (url.contains(htmlOptions.getOpenUrl())) {
+          dialogView.setVisibility(View.VISIBLE);
+          if (activity != null && !activity.isFinishing()) {
+            currentDialog.show();
+          }
+          return true;
+        }
+
+        // Close URL event.
+        if (url.contains(htmlOptions.getCloseUrl())) {
+          cancel();
+          String queryComponentsFromUrl = queryComponentsFromUrl(url, "result");
+          if (!TextUtils.isEmpty(queryComponentsFromUrl)) {
+            Leanplum.track(queryComponentsFromUrl);
+          }
+          return true;
+        }
+
+        // Track URL event.
+        if (url.contains(htmlOptions.getTrackUrl())) {
+          String eventName = queryComponentsFromUrl(url, "event");
+          if (!TextUtils.isEmpty(eventName)) {
+            Double value = Double.parseDouble(queryComponentsFromUrl(url, "value"));
+            String info = queryComponentsFromUrl(url, "info");
+            Map<String, Object> paramsMap = null;
+
+            try {
+              paramsMap = ActionContext.mapFromJson(new JSONObject(queryComponentsFromUrl(url,
+                  "parameters")));
+            } catch (Exception ignored) {
+            }
+
+            if (queryComponentsFromUrl(url, "isMessageEvent").equals("true")) {
+              ActionContext actionContext = htmlOptions.getActionContext();
+              actionContext.trackMessageEvent(eventName, value, info, paramsMap);
+            } else {
+              Leanplum.track(eventName, value, info, paramsMap);
+            }
+          }
+          return true;
+        }
+
+        // Action URL or track action URL event.
+        if (url.contains(htmlOptions.getActionUrl()) ||
+            url.contains(htmlOptions.getTrackActionUrl())) {
+          cancel();
+          String queryComponentsFromUrl = queryComponentsFromUrl(url, "action");
+          try {
+            queryComponentsFromUrl = URLDecoder.decode(queryComponentsFromUrl, "UTF-8");
+          } catch (UnsupportedEncodingException ignored) {
+          }
+
+          ActionContext actionContext = htmlOptions.getActionContext();
+          if (!TextUtils.isEmpty(queryComponentsFromUrl) && actionContext != null) {
+            if (url.contains(htmlOptions.getActionUrl())) {
+              actionContext.runActionNamed(queryComponentsFromUrl);
+            } else {
+              actionContext.runTrackedActionNamed(queryComponentsFromUrl);
+            }
+          }
+          return true;
+        }
+
+        return false;
+      }
+    });
+    String html = htmlOptions.getHtmlTemplate();
+
+    webView.loadDataWithBaseURL(null, html, "text/html", "UTF-8", null);
+
+    return webView;
+  }
+
+  /**
+   * Get query components from URL.
+   *
+   * @param url URL string.
+   * @param components Name of components.
+   * @return String String with query components.
+   */
+  private String queryComponentsFromUrl(String url, String components) {
+    String componentsFromUrl = "";
+    String[] urlComponents = url.split("\\?");
+    if (urlComponents.length > 1) {
+      String queryString = urlComponents[1];
+      String[] parameters = queryString.split("&");
+      for (String parameter : parameters) {
+        String[] parameterComponents = parameter.split("=");
+        if (parameterComponents.length > 1 && parameterComponents[0].equals(components)) {
+          componentsFromUrl = parameterComponents[1];
+        }
+      }
+    }
+    try {
+      componentsFromUrl = URLDecoder.decode(componentsFromUrl, "UTF-8");
+    } catch (Exception ignored) {
+    }
+    return componentsFromUrl;
+  }
+
+  @Override
+  public void onWindowFocusChanged(boolean hasFocus) {
+    try {
+      if (webView != null) {
+        if (hasFocus) {
+          webView.onResume();
+        } else {
+          webView.onPause();
+        }
+      }
+    } catch (Throwable ignore) {
+    }
+    super.onWindowFocusChanged(hasFocus);
+  }
+
+  @Override
+  protected void onFadeOutAnimationEnded() {
+    super.onFadeOutAnimationEnded();
+
+    Handler handler = new Handler();
+    handler.postDelayed(new Runnable() {
+      @Override
+      public void run() {
+        if (webView != null) {
+          webView.stopLoading();
+          webView.loadUrl("");
+          if (dialogView != null) {
+            dialogView.removeAllViews();
+          }
+          webView.removeAllViews();
+          webView.destroy();
+        }
+      }
+    }, 10);
   }
 
   @Override
   public boolean dispatchTouchEvent(@NonNull MotionEvent ev) {
     if (!htmlOptions.isFullScreen()) {
-      if (isBannerWithTapOutsideFalse(htmlOptions)) {
+      if (htmlOptions.isBannerWithTapOutsideFalse()) {
         return super.dispatchTouchEvent(ev);
       }
 
@@ -69,7 +392,7 @@ public class HTMLTemplate extends BaseMessageDialog {
       int htmlYOffset = htmlOptions.getHtmlYOffset(activity);
       int top;
       int bottom;
-      if (MessageTemplates.Args.HTML_ALIGN_BOTTOM.equals(htmlOptions.getHtmlAlign())) {
+      if (htmlOptions.isHtmlAlignBottom()) {
         top = size.y - height - statusBarHeight - htmlYOffset;
         bottom = size.y - htmlYOffset - statusBarHeight;
       } else {
